moved the dataview inside the ShortGrainDropper class
